# packages/contracts/generated/spring/src/main/java/tv/lumo/api/generated/api/AccountApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.14.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package tv.lumo.api.generated.api;

import tv.lumo.api.generated.model.DeviceList;
import tv.lumo.api.generated.model.Entitlement;
import tv.lumo.api.generated.model.Problem;
import java.util.UUID;
import tv.lumo.api.generated.model.UpdateUserRequest;
import tv.lumo.api.generated.model.User;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.lang.Nullable;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.14.0")
@Validated
public interface AccountApi {

    /**
     * GET /me : The authenticated user
     *
     * @return The user. (status code 200)
     *         or Missing, malformed or expired access token (&#x60;UNAUTHENTICATED&#x60;, &#x60;ACCESS_TOKEN_EXPIRED&#x60;). On &#x60;ACCESS_TOKEN_EXPIRED&#x60; the client refreshes once and replays the request.  (status code 401)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/me",
        produces = { "application/json", "application/problem+json" }
    )
    
    ResponseEntity<User> getCurrentUser(
        
    );


    /**
     * GET /me/entitlement : The user&#39;s access rights
     * **The only source of truth for access rights** (ADR 0003).  A client never asks a store whether the user is premium. Both Android listings and the web read this single endpoint. In v1 the table is fed by Stripe webhooks; in v2 Play RTDN will write into the same table, with no client-side change.  A user with no subscription has a &#x60;FREE&#x60; / &#x60;ACTIVE&#x60; entitlement; this endpoint never returns &#x60;404&#x60;. 
     *
     * @return The user&#39;s single active entitlement. (status code 200)
     *         or Missing, malformed or expired access token (&#x60;UNAUTHENTICATED&#x60;, &#x60;ACCESS_TOKEN_EXPIRED&#x60;). On &#x60;ACCESS_TOKEN_EXPIRED&#x60; the client refreshes once and replays the request.  (status code 401)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/me/entitlement",
        produces = { "application/json", "application/problem+json" }
    )
    
    ResponseEntity<Entitlement> getEntitlement(
        
    );


    /**
     * GET /me/devices : Devices linked to the account
     *
     * @return The account&#39;s devices, most recently seen first. (status code 200)
     *         or Missing, malformed or expired access token (&#x60;UNAUTHENTICATED&#x60;, &#x60;ACCESS_TOKEN_EXPIRED&#x60;). On &#x60;ACCESS_TOKEN_EXPIRED&#x60; the client refreshes once and replays the request.  (status code 401)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/me/devices",
        produces = { "application/json", "application/problem+json" }
    )
    
    ResponseEntity<DeviceList> listDevices(
        
    );


    /**
     * DELETE /me/devices/{id} : Unlink a device and revoke its tokens
     * Revokes the device&#39;s entire refresh-token chain. That device is signed out at its next call. Removing the calling device is allowed and is equivalent to signing out. 
     *
     * @param id Resource identifier. (required)
     * @return Device unlinked and its tokens revoked. (status code 204)
     *         or Missing, malformed or expired access token (&#x60;UNAUTHENTICATED&#x60;, &#x60;ACCESS_TOKEN_EXPIRED&#x60;). On &#x60;ACCESS_TOKEN_EXPIRED&#x60; the client refreshes once and replays the request.  (status code 401)
     *         or No such device on this account (&#x60;DEVICE_NOT_FOUND&#x60;). (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/me/devices/{id}",
        produces = { "application/problem+json" }
    )
    
    ResponseEntity<Void> revokeDevice(
         @PathVariable("id") UUID id
    );


    /**
     * PATCH /me : Update display name and locale
     * Only &#x60;display_name&#x60; and &#x60;locale&#x60; are mutable here. Email changes require re-verification and are not part of v1. Omitted properties are left unchanged; an explicit &#x60;null&#x60; clears &#x60;display_name&#x60;. 
     *
     * @param updateUserRequest  (required)
     * @return The updated user. (status code 200)
     *         or The request is malformed or fails validation (&#x60;VALIDATION_FAILED&#x60;). (status code 400)
     *         or Missing, malformed or expired access token (&#x60;UNAUTHENTICATED&#x60;, &#x60;ACCESS_TOKEN_EXPIRED&#x60;). On &#x60;ACCESS_TOKEN_EXPIRED&#x60; the client refreshes once and replays the request.  (status code 401)
     */
    @RequestMapping(
        method = RequestMethod.PATCH,
        value = "/me",
        produces = { "application/json", "application/problem+json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<User> updateCurrentUser(
         @Valid @RequestBody UpdateUserRequest updateUserRequest
    );

}