# packages/contracts/generated/spring/src/main/java/tv/lumo/api/generated/model/EntitlementStatus.java
package tv.lumo.api.generated.model;

import java.net.URI;
import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonValue;
import java.time.OffsetDateTime;
import jakarta.validation.Valid;
import jakarta.validation.constraints.*;


import java.util.*;
import jakarta.annotation.Generated;

import com.fasterxml.jackson.annotation.JsonCreator;
import com.fasterxml.jackson.annotation.JsonValue;

/**
 * State of the entitlement. Only `ACTIVE` grants premium features.
 */

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.14.0")
public enum EntitlementStatus {
  
  ACTIVE("ACTIVE"),
  
  PAST_DUE("PAST_DUE"),
  
  CANCELED("CANCELED"),
  
  EXPIRED("EXPIRED");

  private final String value;

  EntitlementStatus(String value) {
    this.value = value;
  }

  @JsonValue
  public String getValue() {
    return value;
  }

  @Override
  public String toString() {
    return String.valueOf(value);
  }

  @JsonCreator
  public static EntitlementStatus fromValue(String value) {
    for (EntitlementStatus b : EntitlementStatus.values()) {
      if (b.value.equals(value)) {
        return b;
      }
    }
    throw new IllegalArgumentException("Unexpected value '" + value + "'");
  }
}