# packages/contracts/openapi.yaml
openapi: 3.0.3
...
info:
  title: Lumo TV API
  version: 1.0.0
  description: |
    Backend for the Lumo TV IPTV player (`lumo-api`).

    Lumo hosts, provides and resells **no content**. The user brings their own
    source — an M3U playlist or an Xtream Codes account. This API manipulates
    metadata only: catalogue, guide, favourites, playback progress, access
    rights. **Media never transits through this infrastructure**: the player
    opens the stream directly against the user's own IPTV server.

    ## Authentication

    All operations require `Authorization: Bearer <access_token>` unless the
    operation declares `security: []`. Access tokens are short-lived JWTs
    (15 minutes). Refresh tokens are opaque, rotated on every use, and bound to
    a `device`. Presenting an already-consumed refresh token revokes the whole
    token chain for that device — that is the theft-detection mechanism.

    ## Errors

    Every error response is RFC 7807 `application/problem+json` and carries a
    stable, machine-readable `code`. Clients switch on `code` and render their
    own localised message (FR/EN). `detail` is meant for logs and is never
    displayed raw to a user.

    ## Asynchronous ingestion

    Parsing is always server-side. `POST /sources` performs a fast synchronous
    reachability and credential check, then returns `202` with the source in
    `PENDING`; the full catalogue ingestion runs in the background. Clients poll
...
    blocks an HTTP request.
  contact:
...
servers:
...
security:
  - bearerAuth: []

# ==============================================================================
# PATHS
# ==============================================================================
paths:
...
  /me:
    get:
      tags: [account]
      operationId: getCurrentUser
      summary: The authenticated user
      responses:
        '200':
          description: The user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401': { $ref: '#/components/responses/Unauthorized' }
    patch:
      tags: [account]
      operationId: updateCurrentUser
      summary: Update display name and locale
      description: |
        Only `display_name` and `locale` are mutable here. Email changes require
        re-verification and are not part of v1. Omitted properties are left
        unchanged; an explicit `null` clears `display_name`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: The updated user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /me/entitlement:
    get:
      tags: [account]
      operationId: getEntitlement
      summary: The user's access rights
      description: |
        **The only source of truth for access rights** (ADR 0003).

        A client never asks a store whether the user is premium. Both Android
        listings and the web read this single endpoint. In v1 the table is fed by
        Stripe webhooks; in v2 Play RTDN will write into the same table, with no
        client-side change.

        A user with no subscription has a `FREE` / `ACTIVE` entitlement; this
        endpoint never returns `404`.
      responses:
        '200':
          description: The user's single active entitlement.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entitlement'
        '401': { $ref: '#/components/responses/Unauthorized' }
  /me/devices:
    get:
      tags: [account]
      operationId: listDevices
      summary: Devices linked to the account
      responses:
        '200':
          description: The account's devices, most recently seen first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceList'
        '401': { $ref: '#/components/responses/Unauthorized' }
  /me/devices/{id}:
    delete:
      tags: [account]
      operationId: revokeDevice
      summary: Unlink a device and revoke its tokens
      description: |
        Revokes the device's entire refresh-token chain. That device is signed
        out at its next call. Removing the calling device is allowed and is
        equivalent to signing out.
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '204':
          description: Device unlinked and its tokens revoked.
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: No such device on this account (`DEVICE_NOT_FOUND`).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }

  # ------------------------------------------------------------- sources ------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Short-lived access token (15 minutes) obtained from `/auth/login`,
        `/auth/register`, `/auth/oauth/google`, `/auth/refresh` or
        `/auth/device/token`.

  # ---------------------------------------------------------- parameters ------
  parameters:
    PathId:
      name: id
      in: path
      required: true
      description: Resource identifier.
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: The request is malformed or fails validation (`VALIDATION_FAILED`).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: |
        Missing, malformed or expired access token (`UNAUTHENTICATED`,
        `ACCESS_TOKEN_EXPIRED`). On `ACCESS_TOKEN_EXPIRED` the client refreshes
        once and replays the request.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    Problem:
      type: object
      description: |
        RFC 7807 problem detail. Every error response in this API uses this
        shape with `Content-Type: application/problem+json`.

        Clients switch on `code`, never on `title` or `detail`. `detail` is
        written for logs and support, is not localised, and is never displayed
        raw to a user.
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri
          description: Stable URI identifying the problem type.
          example: https://lumo.tv/errors/source-auth-failed
        title:
          type: string
          description: Short, human-readable, English, non-localised summary.
          example: Source authentication failed
        status:
          type: integer
          format: int32
          minimum: 100
          maximum: 599
          description: HTTP status code, repeated for convenience.
          example: 422
        code:
          allOf:
            - $ref: '#/components/schemas/ErrorCode'
          description: |
            Stable machine-readable code. **This is the only field a client is
            allowed to branch on.** Clients must handle an unrecognised value
            gracefully rather than crash: new codes may be added within v1.
        detail:
          type: string
          nullable: true
          description: |
            Diagnostic explanation for logs and support. Never rendered raw in
            the UI. Never contains a stream URL, an Xtream password or a token.
          example: Connection timed out after 10s
        instance:
          type: string
          format: uri-reference
          nullable: true
          description: Path of the request that produced this problem.
          example: /v1/sources
        errors:
          type: array
          nullable: true
          description: Per-field details, present on validation failures only.
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      description: One field-level validation failure.
      required: [field, code]
      properties:
        field:
          type: string
          description: JSON pointer to the offending property, relative to the request body.
          example: /password
        code:
          type: string
          description: Stable machine-readable reason.
          example: TOO_SHORT
        detail:
          type: string
          nullable: true
          description: Diagnostic explanation. Not localised, not for display.
...
    ErrorCode:
      type: string
      description: |
        Every stable error code this API emits, ingestion codes included.

        Grouped by concern:

        - **Generic** — `VALIDATION_FAILED`, `UNAUTHENTICATED`,
          `ACCESS_TOKEN_EXPIRED`, `FORBIDDEN`, `NOT_FOUND`, `CONFLICT`,
          `RATE_LIMITED`, `INTERNAL_ERROR`.
        - **Account and sign-in** — `EMAIL_ALREADY_REGISTERED`,
          `INVALID_CREDENTIALS`, `PASSWORD_TOO_WEAK`, `OAUTH_TOKEN_INVALID`,
          `VERIFICATION_TOKEN_INVALID`, `VERIFICATION_TOKEN_EXPIRED`,
          `RESET_TOKEN_INVALID`, `RESET_TOKEN_EXPIRED`.
        - **Sessions** — `REFRESH_TOKEN_INVALID`, `REFRESH_TOKEN_REUSED`,
          `DEVICE_NOT_FOUND`.
        - **TV activation (RFC 8628)** — `AUTHORIZATION_PENDING`, `SLOW_DOWN`,
          `ACCESS_DENIED`, `EXPIRED_TOKEN`, `DEVICE_CODE_NOT_FOUND`,
          `DEVICE_CODE_EXPIRED`, `DEVICE_CODE_ALREADY_USED`.
        - **Sources and ingestion** — `SOURCE_NOT_FOUND`, `SOURCE_NOT_READY`,
          `SOURCE_SYNC_IN_PROGRESS`, `SOURCE_SYNC_RATE_LIMITED`, plus every
          `IngestionErrorCode`.
        - **Catalogue and user data** — `CHANNEL_NOT_FOUND`,
          `FAVORITE_NOT_FOUND`, `FAVORITE_ALREADY_EXISTS`,
          `FAVORITE_GROUP_NOT_FOUND`, `FAVORITE_GROUP_ALREADY_EXISTS`.
      enum:
        - VALIDATION_FAILED
        - UNAUTHENTICATED
        - ACCESS_TOKEN_EXPIRED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - RATE_LIMITED
        - INTERNAL_ERROR
        - EMAIL_ALREADY_REGISTERED
        - INVALID_CREDENTIALS
        - PASSWORD_TOO_WEAK
        - OAUTH_TOKEN_INVALID
        - VERIFICATION_TOKEN_INVALID
        - VERIFICATION_TOKEN_EXPIRED
        - RESET_TOKEN_INVALID
        - RESET_TOKEN_EXPIRED
        - REFRESH_TOKEN_INVALID
        - REFRESH_TOKEN_REUSED
        - DEVICE_NOT_FOUND
        - AUTHORIZATION_PENDING
        - SLOW_DOWN
        - ACCESS_DENIED
        - EXPIRED_TOKEN
        - DEVICE_CODE_NOT_FOUND
        - DEVICE_CODE_EXPIRED
        - DEVICE_CODE_ALREADY_USED
        - SOURCE_NOT_FOUND
        - SOURCE_NOT_READY
        - SOURCE_SYNC_IN_PROGRESS
        - SOURCE_SYNC_RATE_LIMITED
        - SOURCE_UNREACHABLE
        - SOURCE_AUTH_FAILED
        - SOURCE_EXPIRED
        - SOURCE_MAX_CONNECTIONS
        - SOURCE_INVALID_FORMAT
        - SOURCE_EMPTY
        - SOURCE_TOO_LARGE
        - CHANNEL_NOT_FOUND
        - FAVORITE_NOT_FOUND
        - FAVORITE_ALREADY_EXISTS
        - FAVORITE_GROUP_NOT_FOUND
        - FAVORITE_GROUP_ALREADY_EXISTS

    # ---- enums -------------------------------------------------------------
    Locale:
      type: string
      description: UI language. FR and EN are supported from the first screen.
      enum: [fr, en]
    Platform:
      type: string
      description: The kind of installation a `device` represents.
      enum: [ANDROID_MOBILE, ANDROID_TV, WEB]
...
    Plan:
      type: string
      description: Subscription tier of an `entitlement`.
      enum: [FREE, PREMIUM]
    EntitlementStatus:
      type: string
      description: State of the entitlement. Only `ACTIVE` grants premium features.
      enum: [ACTIVE, PAST_DUE, CANCELED, EXPIRED]
    EntitlementProvider:
      type: string
      description: |
        What fed the entitlement. `STRIPE` in v1, `PLAY` reserved for v2 RTDN,
        `MANUAL` for support-granted access.
      enum: [STRIPE, PLAY, MANUAL]
...
    User:
      type: object
      description: |
        The account. `password_hash` and `oauth_identity` rows are internal and
        appear in no response.
      required: [id, email, locale, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
          nullable: true
        locale:
          $ref: '#/components/schemas/Locale'
        email_verified_at:
          type: string
          format: date-time
          nullable: true
          description: Null until the verification link is followed.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UpdateUserRequest:
      type: object
      description: Omitted properties are left unchanged.
      properties:
        display_name:
          type: string
          nullable: true
          maxLength: 100
          description: An explicit `null` clears it.
        locale:
          $ref: '#/components/schemas/Locale'
    Device:
      type: object
      description: One installation linked to the account.
      required: [id, platform, created_at]
      properties:
        id:
          type: string
          format: uuid
        platform:
          $ref: '#/components/schemas/Platform'
        name:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
        app_version:
          type: string
          nullable: true
        last_seen_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    DeviceList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Device'
    Entitlement:
      type: object
      description: |
        The user's access rights, computed server-side. `provider_ref` is a
        billing-internal reference and is intentionally not exposed.
      required: [plan, status, provider, updated_at]
      properties:
        plan:
          $ref: '#/components/schemas/Plan'
        status:
          $ref: '#/components/schemas/EntitlementStatus'
        provider:
          $ref: '#/components/schemas/EntitlementProvider'
        current_period_end:
          type: string
          format: date-time
          nullable: true
          description: End of the paid period. Null on `FREE`.
        updated_at:
          type: string
          format: date-time

    # ---- sources -----------------------------------------------------------
...